requirements.txt
numpy
pandas
pycoingecko
python-dateutil
requests